feat(): add red habitat banner listing variation for habitat banner refs#301670
