+++ src/components/manage/Blocks/HabitatsBanner/index.js
@@ -1,6 +1,8 @@
 import worldSVG from '@plone/volto/icons/world.svg';
 import Edit from './Edit';
 import View from './View';
+import DefaultVariation from './variations/Default';
+import RedHabitatsVariation from './variations/RedHabitats';
 
 export default function applyConfig(config) {
   config.blocks.blocksConfig.habitats_banner = {
@@ -19,6 +21,19 @@ export default function applyConfig(config) {
       view: [],
     },
     blockHasOwnFocusManagement: false,
+    variations: [
+      {
+        id: 'default',
+        title: 'Default',
+        isDefault: true,
+        template: DefaultVariation,
+      },
+      {
+        id: 'redHabitats',
+        title: 'Red Habitats',
+        template: RedHabitatsVariation,
+      },
+    ],
   };
   return config;
 }
