+++ src/components/manage/Blocks/HabitatsBanner/variations/style.less
